SCALAPACK: Fix added in Test application for few failing test cases with Flang compiler.
Test suite ALL PASS on Ubuntu & GCC with the macro 'F2C_COMPLEX' disabled.
Test suite ALL PASS on Ubuntu & Flang with the macro 'F2C_COMPLEX' enabled.

AMD-Internal: [SWLCSG-124] [CPUPL-566]

+++ TESTING/LIN/pmatgeninc.f
@@ -55,7 +55,12 @@
       PARAMETER        ( IPOW15=2**15, IPOW16=2**16, IPOW30=2**30 )
 *     ..
 *     .. Local Scalars ..
-      INTEGER            KT, LT
+
+*     Declare as 8 byte Integer to avoid overflow during
+*     multiply operation.
+* 
+      INTEGER(KIND=8)     KT, LT
+
 *     ..
 *     .. Intrinsic Functions ..
       INTRINSIC          MOD
